figure update
